Match-id-4c43c1c742f32bf4f3949b56d3c2b0a67a303c6e

+++ edge-manager/pkg/appmanager/app_types.go
@@ -12,7 +12,7 @@ type Container struct {
 	CpuLimit     *float64        `json:"cpuLimit,omitempty"`
 	MemRequest   int64           `json:"memRequest"`
 	MemLimit     *int64          `json:"memLimit,omitempty"`
-	Npu          *float64        `json:"npu,omitempty"`
+	Npu          *int64          `json:"npu,omitempty"`
 	Command      []string        `json:"command"`
 	Args         []string        `json:"args"`
 	Env          []EnvVar        `json:"env"`
